refactor global styles and add main style to MovieDetailPage
