+++ project-movie-app/src/styles/GlobalStyle.ts
